requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
gunicorn==19.7.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==3.3.1
django-filter==2.2.0
